refactor move installable utils
